backend/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.9
django-admin-rangefilter==0.8.2
django-adminlte-3==0.1.6
django-autocomplete-light==3.8.2
django-baton==2.2.3
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-daterangepicker==1.0.6
django-bootstrap3==21.1
django-bootstrap3-daterangepicker==1.0.0b5
django-bootstrap4==21.1
django-cors-headers==3.5.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-filter==21.1
django-forms-bootstrap==3.1.0
django-guardian==2.4.0
django-import-export==2.6.1
django-jazzmin==2.4.8
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.8.0
drf-yasg==1.17.1
et-xmlfile==1.1.0
gunicorn==20.0.4
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
MarkupPy==1.14
MarkupSafe==2.0.1
mysqlclient==2.0.3
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.9
packaging==21.0
Pillow==8.4.0
PyJWT==2.3.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
simplejson==3.17.5
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tablib==3.0.0
uritemplate==4.0.0
urllib3==1.26.7
xlrd==2.0.1
xlwt==1.3.0